+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [3.1.0] (IN PROGRESS)
 
-* Change selected parton background color to increase color contrast. Refs UICHKOUT-603. 
+* Change selected parton background color to increase color contrast. Refs UICHKOUT-603.
 * Fix bug preventing continuation after cancelling checkout notes or multipiece modal. Fixes UICHKOUT-610, UICHKOUT-611.
+* Show confirmation modal when item with withdrawn status is scanned. Refs UICHKOUT-605.
 
 ## [3.0.0](https://github.com/folio-org/ui-checkout/tree/v3.0.0) (2020-03-16)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v2.0.0...v3.0.0)