+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-checkout
 
+## [4.0.0] (IN PROGRESS)
+
+* Upgrade to `stripes` `4.0`, `react-intl` `4.5`. Refs STRIPES-672.
+
 ## [3.1.0] (IN PROGRESS)
 
 * Change selected parton background color to increase color contrast. Refs UICHKOUT-603.